[Console] removed proxy
